Fix concurrency issue on deploy action
This occurs when there are too many PRs merged in a 10-minute
window. To avoid this, we set a concurrency group per branch
(in this case only develop and master apply). I feel like that
cancelling the previous deploy actions makes the merged PR
less semantic because we technically loose track of where the
hotfixes happened.